+++ src/Network/Network/ChangeLog.md
@@ -50,6 +50,7 @@
         - Set-AzLoadBalancerFrontendIpConfig
 * Application Gateway New-AzApplicationGatewayProbeConfig command update for supporting custom port in Probe
     - Updated New-AzApplicationGatewayProbeConfig: Added optional parameter Port which is used for probing backend server. This parameter is applicable for Standard_V2 and WAF_V2 SKU.
+* Updated Get-AzVirtualNetworkSubnetConfig: Added parameter set to get by subnet resource id.
 
 ## Version 1.11.0
 * Added `RoutingPreference` to public ip tags